feat: I create the front-end
