+++ src/app/home/page.jsx
@@ -8,33 +8,6 @@ import { FaBell } from "react-icons/fa";
 import { IoPeople } from "react-icons/io5";
 
 const Home = () => {
-  const [isRaceActive, setIsRaceActive] = useState(false);
-  const [driverInfo, setDriverInfo] = useState({
-    name: "",
-    license: "",
-  });
-  const [startTime, setStartTime] = useState(null);
-  const [endTime, setEndTime] = useState(null);
-
-  const handleStartRace = () => {
-    setStartTime(new Date());
-    setIsRaceActive(true);
-    setEndTime(null);
-  };
-
-  const handleEndRace = () => {
-    setEndTime(new Date());
-    setIsRaceActive(false);
-  };
-
-  const handleInputChange = (e) => {
-    const { name, value } = e.target;
-    setDriverInfo((prevInfo) => ({
-      ...prevInfo,
-      [name]: value,
-    }));
-  };
-
   return (
     <>
       <Header />
@@ -97,69 +70,8 @@ const Home = () => {
           </ul>
           <ul className="flex items-center space-x-2 bg-graciosa p-2 rounded-lg group">
             <IoPeople className="text-2xl text-white" />
-            <a href="" className="text-white text-lg">Novos colaboradores</a>
+            <a href="/travel" className="text-white text-lg">Viagem</a>
           </ul>
-        </div>
-      </div>
-
-      {/* Formulário e Simulação de Corrida */}
-      <div className="flex justify-center">
-        <div className="w-3/5 bg-graciosa p-6 rounded-lg shadow-md">
-          {!isRaceActive ? (
-            <div>
-              <h2 className="text-white text-2xl font-bold mb-4">Iniciar Corrida</h2>
-              <form className="mb-4">
-                <label className="block mb-2 text-white">
-                  Nome do Motorista:
-                  <input
-
-                    type="text"
-                    name="name"
-                    value={driverInfo.name}
-                    onChange={handleInputChange}
-                    className="border p-2 w-full text-black"
-                  />
-                </label>
-                <label className="block mb-4 text-white">
-                  Numero do Ônibus:
-                  <input
-                    type="text"
-                    name="license"
-                    value={driverInfo.license}
-                    onChange={handleInputChange}
-                    className="border p-2 w-full text-black"
-                  />
-                </label>
-                <button
-                  type="button"
-                  onClick={handleStartRace}
-                  className="bg-blue-500 text-white p-2 rounded"
-                >
-                  Iniciar Corrida
-                </button>
-              </form>
-            </div>
-          ) : (
-            <div>
-              <h2 className="text-white text-2xl font-bold mb-4">Corrida em Andamento</h2>
-              <p className="text-white"><strong>Motorista:</strong> {driverInfo.name}</p>
-              <p className="text-white"><strong>Carteira de Habilitação:</strong> {driverInfo.license}</p>
-              <p className="text-white"><strong>Início:</strong> {startTime && startTime.toLocaleString()}</p>
-              <button
-                onClick={handleEndRace}
-                className="bg-red-500 text-white p-2 mt-4 rounded"
-              >
-                Finalizar Corrida
-              </button>
-            </div>
-          )}
-
-          {endTime && (
-            <div className="mt-6">
-              <h2 className="text-white text-xl font-bold">Corrida Finalizada</h2>
-              <p className="text-white"><strong>Fim:</strong> {endTime.toLocaleString()}</p>
-            </div>
-          )}
         </div>
       </div>
     </>
